requirements.txt
python-telegram-bot
requests
requests-oauthlib
python-dotenv